ml-kem: extract `traits` module
Common convention within several of our crates to place trait
definitions inside a `traits` module (as opposed to, in this case,
the toplevel lib.rs)
