+++ ml-kem/src/lib.rs
@@ -65,13 +65,14 @@ pub mod kem;
 /// Section 7. Parameter Sets
 mod param;
 
-use ::kem::{Decapsulate, Encapsulate};
+/// Trait definitions
+mod traits;
+
 use core::fmt::Debug;
 use hybrid_array::{
     Array,
     typenum::{U2, U3, U4, U5, U10, U11, U64},
 };
-use rand_core::CryptoRng;
 
 pub use hybrid_array as array;
 
@@ -79,6 +80,7 @@ pub use hybrid_array as array;
 pub use util::B32;
 
 pub use param::{ArraySize, ParameterSet};
+pub use traits::*;
 
 #[cfg(feature = "pkcs8")]
 pub use pkcs8::{self, AssociatedOid};
@@ -86,77 +88,6 @@ pub use pkcs8::{self, AssociatedOid};
 /// ML-KEM seeds are decapsulation (private) keys, which are consistently 64-bytes across all
 /// security levels, and are the preferred serialization for representing such keys.
 pub type Seed = Array<u8, U64>;
-
-/// An object that knows what size it is
-pub trait EncodedSizeUser {
-    /// The size of an encoded object
-    type EncodedSize: ArraySize;
-
-    /// Parse an object from its encoded form
-    fn from_bytes(enc: &Encoded<Self>) -> Self;
-
-    /// Serialize an object to its encoded form
-    fn as_bytes(&self) -> Encoded<Self>;
-}
-
-/// A byte array encoding a value the indicated size
-pub type Encoded<T> = Array<u8, <T as EncodedSizeUser>::EncodedSize>;
-
-/// A value that can be encapsulated to.  Note that this interface is not safe: In order for the
-/// KEM to be secure, the `m` input must be randomly generated.
-#[cfg(feature = "deterministic")]
-pub trait EncapsulateDeterministic<EK, SS> {
-    /// Encapsulation error
-    type Error: Debug;
-
-    /// Encapsulates a fresh shared secret.
-    ///
-    /// # Errors
-    ///
-    /// Will vary depending on the underlying implementation.
-    fn encapsulate_deterministic(&self, m: &B32) -> Result<(EK, SS), Self::Error>;
-}
-
-/// A generic interface to a Key Encapsulation Method
-pub trait KemCore: Clone {
-    /// The size of a shared key generated by this KEM
-    type SharedKeySize: ArraySize;
-
-    /// The size of a ciphertext encapsulating a shared key
-    type CiphertextSize: ArraySize;
-
-    /// A decapsulation key for this KEM
-    type DecapsulationKey: Decapsulate<Ciphertext<Self>, SharedKey<Self>>
-        + EncodedSizeUser
-        + Debug
-        + PartialEq;
-
-    /// An encapsulation key for this KEM
-    #[cfg(not(feature = "deterministic"))]
-    type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>>
-        + EncodedSizeUser
-        + Clone
-        + Debug
-        + PartialEq;
-
-    /// An encapsulation key for this KEM
-    #[cfg(feature = "deterministic")]
-    type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>>
-        + EncapsulateDeterministic<Ciphertext<Self>, SharedKey<Self>>
-        + EncodedSizeUser
-        + Clone
-        + Debug
-        + PartialEq;
-
-    /// Generate a new (decapsulation, encapsulation) key pair
-    fn generate<R: CryptoRng + ?Sized>(
-        rng: &mut R,
-    ) -> (Self::DecapsulationKey, Self::EncapsulationKey);
-
-    /// Generate a new (decapsulation, encapsulation) key pair deterministically
-    #[cfg(feature = "deterministic")]
-    fn generate_deterministic(d: B32, z: B32) -> (Self::DecapsulationKey, Self::EncapsulationKey);
-}
 
 /// `MlKem512` is the parameter set for security category 1, corresponding to key search on a block
 /// cipher with a 128-bit key.
@@ -266,6 +197,7 @@ pub type MlKem1024 = kem::Kem<MlKem1024Params>;
 #[cfg(test)]
 mod test {
     use super::*;
+    use ::kem::{Decapsulate, Encapsulate};
 
     fn round_trip_test<K>()
     where
