+++ ml-kem/src/traits.rs
@@ -1,0 +1,79 @@
+use crate::{ArraySize, Ciphertext, SharedKey};
+use core::fmt::Debug;
+use hybrid_array::Array;
+use kem::{Decapsulate, Encapsulate};
+use rand_core::CryptoRng;
+
+#[cfg(feature = "deterministic")]
+use crate::util::B32;
+
+/// An object that knows what size it is
+pub trait EncodedSizeUser {
+    /// The size of an encoded object
+    type EncodedSize: ArraySize;
+
+    /// Parse an object from its encoded form
+    fn from_bytes(enc: &Encoded<Self>) -> Self;
+
+    /// Serialize an object to its encoded form
+    fn as_bytes(&self) -> Encoded<Self>;
+}
+
+/// A byte array encoding a value the indicated size
+pub type Encoded<T> = Array<u8, <T as EncodedSizeUser>::EncodedSize>;
+
+/// A value that can be encapsulated to.  Note that this interface is not safe: In order for the
+/// KEM to be secure, the `m` input must be randomly generated.
+#[cfg(feature = "deterministic")]
+pub trait EncapsulateDeterministic<EK, SS> {
+    /// Encapsulation error
+    type Error: Debug;
+
+    /// Encapsulates a fresh shared secret.
+    ///
+    /// # Errors
+    ///
+    /// Will vary depending on the underlying implementation.
+    fn encapsulate_deterministic(&self, m: &B32) -> Result<(EK, SS), Self::Error>;
+}
+
+/// A generic interface to a Key Encapsulation Method
+pub trait KemCore: Clone {
+    /// The size of a shared key generated by this KEM
+    type SharedKeySize: ArraySize;
+
+    /// The size of a ciphertext encapsulating a shared key
+    type CiphertextSize: ArraySize;
+
+    /// A decapsulation key for this KEM
+    type DecapsulationKey: Decapsulate<Ciphertext<Self>, SharedKey<Self>>
+        + EncodedSizeUser
+        + Debug
+        + PartialEq;
+
+    /// An encapsulation key for this KEM
+    #[cfg(not(feature = "deterministic"))]
+    type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>>
+        + EncodedSizeUser
+        + Clone
+        + Debug
+        + PartialEq;
+
+    /// An encapsulation key for this KEM
+    #[cfg(feature = "deterministic")]
+    type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>>
+        + EncapsulateDeterministic<Ciphertext<Self>, SharedKey<Self>>
+        + EncodedSizeUser
+        + Clone
+        + Debug
+        + PartialEq;
+
+    /// Generate a new (decapsulation, encapsulation) key pair
+    fn generate<R: CryptoRng + ?Sized>(
+        rng: &mut R,
+    ) -> (Self::DecapsulationKey, Self::EncapsulationKey);
+
+    /// Generate a new (decapsulation, encapsulation) key pair deterministically
+    #[cfg(feature = "deterministic")]
+    fn generate_deterministic(d: B32, z: B32) -> (Self::DecapsulationKey, Self::EncapsulationKey);
+}
